% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{add_bars_to_condensed_phyloxml}
\alias{add_bars_to_condensed_phyloxml}
\title{Add bars data to a PhyloXML containing unique sequences}
\usage{
add_bars_to_condensed_phyloxml(xml_file, seqs)
}
\arguments{
\item{xml_file}{A path to a PhyloXML file that has been built using only
unique sequences.}

\item{seqs}{A character vector containing non-unique sequences. This function
will use the frequencies of sequences in this vector to add bars to the
corresponding sequences in the PhyloXML. The \code{xml_file} representing the
PhyloXML should have been built from the unique sequences in this vector,
i.e. with \code{unique(seqs)}.}
}
\value{
A path to the PhyloXML file annotated with bars and tooltip data.
}
\description{
An internal function that performs the final step of createing a
"condensed" phylogram. It calculates sequence frequencies and adds the XML
data in order to show bars (representing clonal expansion) and a tooltip on
the radial phylogram.
}
\seealso{
\code{\link{radial_phylo}}
}
\keyword{internal}

