% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onload_checks.R
\name{add_binaries_to_path}
\alias{add_binaries_to_path}
\title{Add HMMER and NCBI BLAST+ binaries to the system path}
\usage{
add_binaries_to_path()
}
\description{
HMMER and BLAST+ binaries are included for multiple platforms.
For HMMER, there are 'windows' and 'unix' binaries. For BLAST+, there are
'windows', 'mac-osx', 'linux x86', and 'linux 'linux ia32' binaries. This
function detects the current operating system and adds the appropriate
binaries to the system \code{PATH} variable if they are not already in
the \code{PATH}.
}
\keyword{internal}

