% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_plotting.R
\name{avg_sil_plot}
\alias{avg_sil_plot}
\title{Plot average silhouette widths for different numbers of clusters.}
\usage{
avg_sil_plot(clust_obj, optimal = FALSE, ...)
}
\arguments{
\item{clust_obj}{A \code{\link{cluster_optimal}} object from which to extract
\code{clust_gap}.}

\item{optimal}{Logical. If \code{TRUE}, the optimal number of clusters as
extracted from \code{clust_obj}, based on average silhouette width, is
circled in the plot.}

\item{...}{Further arguments to be passed to the \code{\link{plot}}
  function (besides \code{xlab}, \code{ylab}, \code{main}).}
}
\description{
This function plots the average silhouette widths for all numbers of
clusters present in \code{clust_obj}. It provides a wrapper on
\code{\link[graphics]{plot}} by extracting \code{sil_avg} from
\code{clust_obj} and plotting it.
}
\examples{
# First, create a cluster_optimal object
library(datasets)
iris_cluster <- cluster_optimal(iris[, 1:4])
# Second, use object with avg_sil_plot
avg_sil_plot(iris_cluster, optimal = TRUE)
}
\seealso{
\code{\link[graphics]{plot}}, \code{\link{cluster_optimal}},
  \code{\link[cluster]{silhouette}}
}

