% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/onload_checks.R
\name{biopy_version}
\alias{biopy_version}
\title{Get Biopython version}
\usage{
biopy_version()
}
\value{
If Biopython is installed the version number will be returned as a
  \code{numeric} (for example, \code{1.62}). If Biopython is not installed,
  returns \code{NULL}.
}
\description{
Get the major and minor version of the installed Biopython package.
}
\details{
This is an internal function for checking to see whether or not Biopython is
installed on the system during the \code{.onLoad()} hook.
}
\keyword{internal}

