% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plotting.R
\name{boxplot_num_cols}
\alias{boxplot_num_cols}
\title{Calculate axis label font size for boxplot}
\usage{
boxplot_num_cols(num_clust)
}
\arguments{
\item{num_clust}{The num_clust argument from \code{\link{clust_boxplot}}}
}
\value{
An integer value for the number of columns
}
\description{
This is an internal function that returns a number of columns for the
\code{\link{clust_boxplot}} function that should minimize tightness of
plot for high number of clusters.
}
\keyword{internal}

