% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{calculate_arc_size}
\alias{calculate_arc_size}
\title{Calculate the arc size for a radial phylogram}
\usage{
calculate_arc_size(xml_file)
}
\arguments{
\item{xml_file}{A path to a PhyloXML file.}
}
\value{
An integer value representing the arc size for the phylogram
}
\description{
An internal function that takes into account the number of
sequences in a PhyloXML and adjusts the arc size the achieve uniform spacing
between sequences.
}
\seealso{
\code{\link{radial_phylo}}
}
\keyword{internal}

