% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{calculate_canvas_size}
\alias{calculate_canvas_size}
\title{Calculate the canvas size for a radial phylogram}
\usage{
calculate_canvas_size(xml_file, condensed, rings)
}
\arguments{
\item{xml_file}{A path to a PhyloXML file.}

\item{condensed}{\code{TRUE} or \code{FALSE}, depending on whether or not the
PhyloXML represents a condensed phylogram (i.e. a phylogram with only
unique sequences and bars representing clonal expansion).}

\item{rings}{A user-supplied named character vector representing rings to add
to a phylogram.}
}
\value{
An integer value representing both the pixel width and height for the
canvas.
}
\description{
An internal function that takes into account the length of
sequences in a PhyloXML, whether or not the phylogram is condensed, and the
number of rings surrounding the phlogram in order to calculate the optimal
canvas size for drawing the SVG. The calculations tend to err on the side of
creating a larger canvas than a smaller one in order to not cut off parts of
the phylogram when it is drawn.
}
\seealso{
\code{\link{radial_phylo}}
}
\keyword{internal}

