% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/onload_checks.R
\name{check_bio_python}
\alias{check_bio_python}
\title{Check for the existence of Python and Biopython}
\usage{
check_bio_python(level)
}
\arguments{
\item{level}{Either \code{"startup_warn"}, \code{"warn"}, or \code{"stop"},
depending on how you want to handle the non-existence of Python or
Biopython.}
}
\value{
Returns TRUE if the libraries are found, otherwise returns
  \code{NULL}.
}
\description{
Reads the versions of Python and Biopython from this package's
  options and warns the user or stops execution depending on the paramater
  \code{level}.
}
\keyword{internal}

