% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_plotting.R
\name{check_boolean}
\alias{check_boolean}
\title{Check if all columns of data.frame are boolean}
\usage{
check_boolean(d)
}
\arguments{
\item{d}{The data argument from \code{\link{clust_boxplot}}}
}
\value{
A boolean value indicating if all columns are boolean
}
\description{
This is an internal function that returns if \code{d} is boolean for the
\code{\link{clust_boxplot}} function that determines if a barplot should be
used.
}
\keyword{internal}

