% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/onload_checks.R
\name{check_muscle}
\alias{check_muscle}
\title{Check for correct version of MUSCLE}
\usage{
check_muscle(level)
}
\arguments{
\item{level}{Either \code{"startup_warn"}, \code{"warn"}, or \code{"stop"},
depending on how you want to handle the non-existence of the \code{muscle}
package or an outdated version of the package.}
}
\value{
Returns TRUE if the correct version of MUSCLE is found, otherwise
  returns \code{NULL}.
}
\description{
Reads the version of \code{muscle} from this package's
  options and warns the user or stops execution depending on the parameter
  \code{level}.
}
\keyword{internal}

