% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_plotting.R
\name{clusGap_plot}
\alias{clusGap_plot}
\title{Plot of cluster gap statistic}
\usage{
clusGap_plot(clust_obj, optimal = FALSE, ...)
}
\arguments{
\item{clust_obj}{A \code{\link{cluster_optimal}} object from which to extract
\code{clust_gap}.}

\item{optimal}{Logical. If \code{TRUE}, the optimal number of clusters as
extracted from \code{clust_obj}, based on average silhouette width, is
circled in the plot.}

\item{...}{Further arguments to be passed to the \code{\link{plot}}
  function (besides \code{xlab}, \code{ylab}, \code{main}).}
}
\description{
This function plots gap analysis. It provides a wrapper on
\code{\link[graphics]{plot}} by extracting \code{clust_gap}, a
\code{\link[cluster]{clusGap}} object, and leveraging the custom plotting
method that it has.
}
\examples{
# First, create a cluster_optimal object
library(datasets)
iris_cluster <- cluster_optimal(iris[, 1:4])
# Second, use object with clusGap_plot
clusGap_plot(iris_cluster, optimal = TRUE)
}
\seealso{
\code{\link[graphics]{plot}}, \code{\link{cluster_optimal}},
  \code{\link[cluster]{clusGap}}
}

