% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_plotting.R
\name{clust_boxplot}
\alias{clust_boxplot}
\title{Plot cluster membership for each feature.}
\usage{
clust_boxplot(d, clust_obj, num_clust, ...)
}
\arguments{
\item{d}{A numeric matrix of data, or an object that can be coerced to
such a matrix (such as a numeric vector or a data frame with all numeric
columns). Note: This should be the same one used to generate
\code{clust_obj}.}

\item{clust_obj}{A \code{\link{multiClust}} object from which to extract
\code{clust_model} based on the argument \code{num_clust}}

\item{num_clust}{An integer. The desired number of clusters to be used. Note:
This integer should fall within the krange used to generate the
\code{\link{multiClust}} object.}

\item{...}{Further arguments to be passed to the \code{\link{qplot}}
  function (besides \code{xlab}, \code{ylab}).}
}
\description{
This function plots cluster membership for each feature of \code{data} using
box plots. It utlizes \code{facet_wrap} in the \code{\link[ggplot2]{qplot}}
function and the desired \code{clust_model} from the argument
\code{clust_obj} to show the box plots of each feature after using the
\code{\link[reshape2]{melt}} function to get the correct form of \code{data}.
}
\examples{
# First, create a multiClust object
library(datasets)
iris_cluster <- multiClust(iris[, 1:4])
# Second, use object with clust_boxplot
clust_boxplot(iris[, 1:4], iris_cluster, num_clust = 3)
}
\seealso{
\code{\link[ggplot2]{qplot}}, \code{\link{multiClust}},
  \code{\link[reshape2]{melt}}
}

