% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_obj.R
\name{cluster_obj}
\alias{cluster_obj}
\title{Cluster with kmeans and find optimal k}
\usage{
cluster_obj(data, krange = 2:10, iter.max = 100, runs = 100, ...)
}
\arguments{
\item{data}{A numeric matrix of data, or an object that can be coerced to
such a matrix (such as a numeric vector or a data frame with all numeric
columns).}

\item{krange}{An integer vector. Numbers of clusters which are to be compared
by the average silhouette width criterion.}

\item{iter.max}{An integer. The maximum number of iterations allowed.}

\item{runs}{An integer. Number of starts of the k-means algorithm.}

\item{...}{further arguments to be passed to \code{\link{kmeans}}.}
}
\value{
\code{cluster_obj} returns an object of class "cluster_obj" that can
  be used for multiple plots. It is a list with at least the following
  components:
  \code{km_obj} A list of \code{\link{kmeans}} objects for each number of
    cluster centers requested.
  \code{sil_avg} A list of average silhouette scores for each number of
    cluster centers requested.
  \code{num_clusters} A list of the each number of clusters used.
  \code{sil} A list of \code{\link{silhouette}} objects for each number of
    clusters.
  \code{k_best} The optimal number of clusters based on silhouette score.
}
\description{
This calls the function \code{\link{kmeans}} to perform kmeans clustering,
but initializes multiple times. It chooses the best one for each number of
clusters, and the best number of clusters overall, based on average
silhouette width using the \code{\link{silhouette}} function.
}
\examples{
library(datasets)
iris_cluster <- cluster_obj(iris[, 1:4])
}
\seealso{
\code{\link{kmeans}}, \code{\link{silhouette}}
}

