% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{compute_dist_matrix}
\alias{compute_dist_matrix}
\title{Compute a distance matrix from a MSA}
\usage{
compute_dist_matrix(ms_alignment, seqs)
}
\arguments{
\item{ms_alignment}{A multiple sequence alignment, as a
\emph{MultipleAlignment} class.}

\item{seqs}{A character vector containing sequences, preferably cleaned by
\code{\link{clean_data}}.}
}
\value{
A distance matrix of class \emph{dist}.
}
\description{
An internal function that uses the \emph{seqinr} package to
  compute an identity distance matrix.
}
\keyword{internal}

