% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{convergence}
\alias{convergence}
\title{Create an interactive convergence network diagram}
\usage{
convergence(d, seqs_col = NULL, browser = FALSE, verbose = FALSE)
}
\description{
Creates a JavaScript-based network diagram in RStudio or a browser window.
This visualization can help identify clusters of functionally similar
complementary determining regions (CDRs) of antibody or T-cell receptors.
}
\examples{

}

