% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_plotting.R
\name{convergence_plot}
\alias{convergence_plot}
\title{Create an interactive convergence group network diagram}
\usage{
convergence_plot(convergence_obj, group_num = NULL,
  background_color = "#FFFFFF", node_shape = "ELLIPSE", border_width = 2,
  border_color = "#161616", node_color = "#0B94B1", node_size = 30,
  labels = TRUE, label_vertical_pos = "middle",
  label_horizontal_pos = "center", edge_width = 3, edge_color = "#2D2D2D",
  browser = FALSE, verbose = FALSE)
}
\arguments{
\item{convergence_obj}{An object of class \code{\link{convergenceGroups}},
generated by the function \code{\link{convergence}}.}

\item{group_num}{A row number corresponding to the convergence group to
visualize. Specifically, this is a row number in the
\code{convergence_obj@groups} data frame, which contains all the groups
resulting from the convergence analysis. For more information on selecting
a group, see the \code{\link{convergence}} documentation.}

\item{background_color}{A valid hex color code as a string for the canvas
color.}

\item{node_shape}{A string indicating what shape to use for the nodes. One of
\code{"ELLIPSE"}, \code{"RECTANGLE"}, \code{"TRIANGLE"}, \code{"DIAMOND"},
\code{"HEXAGON"}, \code{"OCTAGON"}, \code{"PARALLELOGRAM"},
\code{"ROUNDRECT"}, \code{"VEE"}.}

\item{border_width}{An integer border width for the nodes.}

\item{border_color}{A valid hex color code as a string for the border color
of the nodes.}

\item{node_color}{A valid hex color code as a string for the node color.}

\item{node_size}{An integer representing node size.}

\item{labels}{A logical indicating whether or not node labels should be
displayed. If \code{TRUE}, labels will be the sequences used for analysis in
the convergence pipeline.}

\item{label_vertical_pos}{A string indicating the vertical position of node
labels. Must be one of \code{"top"}, \code{"middle"}, or \code{"bottom"}.
Ignored if \code{label=FALSE}.}

\item{label_horizontal_pos}{A string indicating the horizontal position of
node labels. Must be one of \code{"left"}, \code{"center"}, or
\code{"right"}. Ignored if \code{label=FALSE}.}

\item{edge_width}{An integer representing the thickness of edges, or the
lines that connect nodes.}

\item{edge_color}{A valid hex color code as a string for the edge lines.}

\item{browser}{\code{TRUE} or \code{FALSE}, depending on whether or not the 
visualization should open in a browser window. Only applicable if using
RStudio, which by default will show the visualization in RStudio's Viewer.
If using R from the command line, the visualization will always open in a
browser.}

\item{verbose}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, the XML file for
Cytoscape is written to a folder in the working directory.}
}
\description{
Creates a JavaScript-based network diagram in RStudio or a
browser window. This visualization plots groups of functionally similar
complementary determining regions (CDRs) of antibody or T-cell receptors.
These groups are created by the \code{\link{convergence}} function.
}
\examples{
data(tcr)  # Packaged data set, a data.frame from a CSV file
tcr_reduced <- tcr[1:100, ]
converged <- convergence(tcr_reduced, seqs_col='seqs')
# Plot the group with the largest size
largest_group <- which.max(converged@groups[, 1])
convergence_plot(converged, row_num=largest_group)
}

