% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{convergence_xml_path}
\alias{convergence_xml_path}
\title{Create a path to a tempory XML file for plotting with Cytoscape}
\usage{
convergence_xml_path()
}
\value{
A filepath that can be used to save a XML.
}
\description{
An internal function that creates a temporary directory to hold
the XML file for a Cytoscape network diagram. The directory created is only
intended to hold the XML, and no other files, because in order for
htmlwidgets to make the XML available in the HTML it copies the entire parent
directory.
}
\keyword{internal}

