% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_plotting.R
\name{convergenceOutput}
\alias{convergenceOutput}
\title{Widget output function for use in Shiny}
\usage{
convergenceOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

