% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{create_tree}
\alias{create_tree}
\title{Create a phylogenetic tree}
\usage{
create_tree(dist_matrix)
}
\arguments{
\item{dist_matrix}{A distance matrix of class \code{dist}, likely created by
the \code{\link{compute_dist_matrix}} function.}
}
\value{
A path to a Newick file representing the phylogenetic tree.
}
\description{
An internal function that uses the \emph{ape} package and the
BIONJ method (an improved NJ, or neighbor joining, method) to create a
phylogenetic tree.
}
\keyword{internal}

