% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{cytoscape_xml}
\alias{cytoscape_xml}
\title{Create a XML file to plot with Cytoscape}
\usage{
cytoscape_xml(d, row_num, verbose, verbose_dir)
}
\value{
A path to the Cytoscape XML file.
}
\description{
An internal function that parses the output of the convergence
tool and creates an XML file to represent a specific row (cluster) in the
file.
}
\keyword{internal}

