% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{fluidigm}
\alias{fluidigm}
\title{Phenotypic Marker Expression data set}
\format{A data frame with 1292 rows and 22 columns:
\describe{
  \item{CCR7}{phenotypic marker}
  \item{CCR8}{phenotypic marker}
  \item{CD127}{phenotypic marker}
  \item{CD25}{phenotypic marker}
  \item{CD27}{phenotypic marker}
  \item{CD28}{phenotypic marker}
  \item{CD38}{phenotypic marker}
  \item{CD45RA}{phenotypic marker}
  \item{CD69}{phenotypic marker}
  \item{CD90}{phenotypic marker}
  \item{FOXP3}{phenotypic marker}
  \item{IFNg}{phenotypic marker}
  \item{IL.10}{phenotypic marker}
  \item{IL.13}{phenotypic marker}
  \item{IL.4}{phenotypic marker}
  \item{IL.5}{phenotypic marker}
  \item{IL17A}{phenotypic marker}
  \item{IL17F}{phenotypic marker}
  \item{IP.10}{phenotypic marker}
  \item{ITGA4}{phenotypic marker}
  \item{ITGB7}{phenotypic marker}
  \item{Ki.67}{phenotypic marker}
}}
\usage{
data(fluidigm)
}
\description{
A data set containing phenotypic markers and their relative
expression levels. Each row corresponds to data for a single cell.
}
\details{
No thresholds have been applied. All data is as measured.
}
\keyword{datasets}

