% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plotting.R
\name{gap_plot}
\alias{gap_plot}
\title{Plot of cluster gap statistic}
\usage{
gap_plot(clust_obj, optimal = FALSE, ...)
}
\arguments{
\item{clust_obj}{A \code{\link{multiClust-class}} object from which to
extract \code{clust_gap}.}

\item{optimal}{Logical. If \code{TRUE}, the optimal number of clusters as
extracted from \code{clust_obj}, based on average silhouette width, is
circled in the plot.}

\item{...}{Further arguments to be passed to the \code{\link{plot}} 
function (besides \code{xlab}, \code{ylab}, \code{main}).}
}
\description{
This function plots gap analysis. It provides a wrapper on 
\code{\link[graphics]{plot}} by extracting \code{clust_gap}, a 
\code{\link[cluster]{clusGap}} object, and leveraging the custom plotting
method that it has.
}
\details{
This function is intended to be used as part of deciding the ideal
  number of clusters to use for analysis. The cluster gap statistic tells
  one the goodness of a given clustering using the gap statistic. Typically,
  one looks for the peak before which the gap rises significantly. See
  \code{\link[cluster]{clusGap}} for further details on its calculation.
  Sometimes, the peak in the gap still does not contain enough clusters to
  understand the data well. This is where other visualizations can be useful.  
  Choosing \code{optimal = TRUE} will circle the optimal number of clusters
  based on average silhouette width. See the \emph{Details} section for the
  \code{\link{multi_clust}} function and view the \emph{TIP} for a suggested
  workflow.
}
\examples{
# First, create a multiClust object
library(datasets)
iris_cluster <- multi_clust(iris[, 1:4])
# Second, use object with gap_plot
gap_plot(iris_cluster, optimal = TRUE)
}
\seealso{
\code{\link{multi_clust}}, \code{\link{wss_plot}}, 
  \code{\link{pca_plot}}, \code{\link{sil_plot}}, \code{\link{avg_sil_plot}},
  \code{\link{clust_boxplot}}
}

