% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onload_checks.R
\name{get_architecture}
\alias{get_architecture}
\title{Get the architecture of the current machine}
\usage{
get_architecture()
}
\value{
Either 'x86_64' or 'ia32', depending on the architecture.
}
\description{
Checks to see if a machine uses a x86_64 or ia32 processor.
}
\keyword{internal}

