% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onload_checks.R
\name{get_os}
\alias{get_os}
\title{Get the OS type}
\usage{
get_os()
}
\value{
Either 'win', 'osx', or 'linux', as a string.
}
\description{
Polls the current system to get the platform type: 'win',
'osx', or 'linux'. The discovery of 'osx' might not be 100%, so if the
platform is not discovered to be either one of "win", "osx", but it is found
to be "unix" then the return OS type will be "linux". If the platform is
not found to be any of the three ("win", "osx", or "unix") then an error
is thrown. Sinced OS X is a Unix variation, the default 'linux'
should generally be applicable even if the platform is OS X but that
couldn't be discovered properly.
}
\keyword{internal}

