% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{index_with_wrap}
\alias{index_with_wrap}
\title{Index a vector and allow the indexation to wrap if necessary}
\usage{
index_with_wrap(index, v)
}
\arguments{
\item{index}{An index to use to subset the vector}

\item{v}{A vector, either named or not named}
}
\value{
A list containing the \code{"name"} and \code{"val"} of the element.
}
\description{
An internal function that allows for indexation of a vector to
wrap back to the beginning of a vector if the index is greater than the
vector's length.
}
\examples{
\dontrun{
color <- c(white="#FFF", black="000", blue="#00F")
index_with_wrap(5, color)
}
}
\keyword{internal}

