% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_validation.R
\name{is_boolean}
\alias{is_boolean}
\title{Check if all columns of data.frame are boolean or binary}
\usage{
is_boolean(d)
}
\arguments{
\item{d}{A data.frame or matrix.}
}
\value{
A logical indicating whether or not \emph{all} columns are boolean or
binary.
}
\description{
This is an internal function that returns \code{TRUE} if \code{d} is boolean
or binary.
}
\keyword{internal}

