% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{msa}
\alias{msa}
\title{Perform multiple sequence alignment on sequences}
\usage{
msa(seqs, verbose, verbose_dir, dedupe_hash = NULL)
}
\arguments{
\item{seqs}{A character vector containing sequences, preferably cleaned by
\code{\link{clean_data}}.}

\item{verbose}{\code{TRUE} or \code{FALSE}, depending on whether or not the
output of the MSA should be printed and the alignment file should be copied
to the \code{verbose_dir}.}

\item{verbose_dir}{A directory to write files to when \code{verbose} is
\code{TRUE}.}

\item{dedupe_hash}{A random string to use to deduplicate sequence names when
writing the MSA to FASTA. This is necessary to prevent Biopython from
complaining about duplicate names when reading the MSA. By passing this value
in, it can later be removed from any subsequent files, such as the final
PhyloXML.}
}
\value{
A named list containing the multiple sequence alignment
\code{"as_string"} (a \code{MultipleAlignment} class) and a path to the MSA
FASTA \code{"file"}.
}
\description{
An internal function that utilizes Bioconductor's MUSCLE for
MSA. Writes the alignment in addition to returning it so the file can be
loaded by Biopython.
}
\keyword{internal}

