% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{multiClust-class}
\alias{multiClust-class}
\title{An S4 class to hold the results of multiple clusterings from estimate k}
\description{
Each instance of this class is generated by
\code{\link{multi_clust}}, which performs clustering on a data set using
a range of k clusters and saves the result of each model to an instance of
this class.
}
\details{
Each instance of this class contains a number of
\code{\link[stats]{kmeans}} objects, the number of which is based on
\code{@num_clust}. For example, if one generates an instance of this class
using \code{obj <- multi_clust(d, krange=1:3)}, each of the three kmeans
objects is saved in a list at \code{obj@clust_model}. Additional information
about each kmeans is contained in \code{@sil_avg}, \code{@num_clust},
\code{@sil}, and \code{@wss}.

Both \code{@clust_gap} and \code{@k_best} only contain single items. Thus,
\code{@clust_gap} can be accessed with the following call:
\code{obj@clust_gap}.
}
\section{Slots}{

\describe{
\item{\code{num_clust}}{A vector of each number of clusters, k, used on the original
data set. The length of this vector is the same as the length of
\code{clust_model}.}

\item{\code{clust_model}}{A list of \code{\link{kmeans}} models. The length of the
list is indicative of the number of clusters, k, run on the original data
set.}

\item{\code{sil_avg}}{A numeric vector of average silhouette scores for each number
of clusters, k, requested. This vector is the same length as
\code{clust_model}, and the index of each average silhouette score
corresponds to its kmeans model in \code{clust_model}.}

\item{\code{sil}}{A list of \code{\link[cluster]{silhouette}} objects - one for each
kmeans model in \code{clust_model}.}

\item{\code{wss}}{A numeric vector of the total within sum of squares for each of the
\code{\link[stats]{kmeans}} objects stored in \code{clust_model}.}

\item{\code{clust_gap}}{A \code{\link[cluster]{clusGap}} object that uses the
highest number of \code{krange} for the \code{K.max} argument. This object
represents a gap analysis based on the number of clusters specified in
\code{num_clust}.}

\item{\code{k_best}}{The suggested optimal number of clusters to use when describing
the original data set.
based on the algorithms used
by \code{\link[NbClust]{NbClust}}. These include: "kl", "ch",
"hartigan", "ccc", "scott", "marriot", "trcovw", "tracew", "friedman",
"rubin", "cindex", "db", "silhouette", "duda", "pseudot2", "beale",
"ratkowsky", "ball", "ptbiserial", "gap", "frey", "mcclain", "gamma",
"gplus", "tau", "dunn", "hubert", "sdindex", "dindex", "sdbw".}
}}
\seealso{
\code{\link{wss_plot}}, \code{\link{gap_plot}}, 
\code{\link{pca_plot}}, \code{\link{sil_plot}}, \code{\link{avg_sil_plot}},
\code{\link{clust_boxplot}}, \code{\link{multi_clust}}
}

