% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multi_clustering.R
\name{multiClust}
\alias{multiClust}
\title{Cluster with kmeans and find optimal k}
\usage{
multiClust(d, krange = 2:10, iter.max = 300, runs = 10,
  method = "kmeans", ...)
}
\arguments{
\item{d}{A numeric matrix of data, or an object that can be coerced to
such a matrix (such as a numeric vector or a data frame with all numeric
columns).}

\item{krange}{An integer vector. Numbers of clusters which are to be compared
by the average silhouette width criterion. Note: \code{krange} should not
include 1 since silhouette scores are not defined there.}

\item{iter.max}{An integer. The maximum number of iterations used by kmeans
to find its centers.}

\item{runs}{An integer. The number of starts of the k-means algorithm.}

\item{method}{A string. Either "kmeans", "hclust", or "both". Currently, it
does nothing, but functionality will be added in the future.}

\item{...}{Further arguments to be passed to \code{\link[stats]{kmeans}}.}
}
\value{
\code{multiClust} returns an object of class "multiClust"
  that can be used for multiple plots. It is a list with at least the
  following components:
  \code{clust_model} A list of \code{\link[stats]{kmeans}} objects for each
    number of cluster centers requested.
  \code{sil_avg} A list of average silhouette scores for each number of
    cluster centers requested.
  \code{num_clust} A list of the each number of clusters used.
  \code{sil} A list of \code{\link[cluster]{silhouette}} objects for each
    number of clusters.
  \code{clust_gap} A \code{\link[cluster]{clusGap}} object that uses the
    highest number of \code{krange} for the \code{K.max} argument.
  \code{k_best} The optimal number of clusters based on silhouette score.
}
\description{
This calls the function \code{\link[stats]{kmeans}} to perform kmeans
clustering, but initializes multiple times. It chooses the best one for each
number of clusters using within sum of squares, and the best number of
clusters overall based on average silhouette width from the
\code{\link[cluster]{silhouette}} function.
}
\examples{
library(datasets)
iris_cluster <- multiClust(iris[, 1:4])
}
\seealso{
\code{\link[stats]{kmeans}}, \code{\link[cluster]{silhouette}},
  \code{\link[cluster]{clusGap}}
}

