% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{newick_to_phyloxml}
\alias{newick_to_phyloxml}
\title{Convert a Newick file to a PhyloXML}
\usage{
newick_to_phyloxml(newick_file, verbose, verbose_dir)
}
\arguments{
\item{newick_file}{A path to the Newick file, likely created by the
\code{\link{create_tree}} function.}
}
\value{
A path to the PhyloXML file.
}
\description{
An internal function that utilizes the Forester Java library to
  perform the conversion. The converted PhyloXML file will be saved to a
  tempory directory.
}
\keyword{internal}

