% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{parse_convergence_groups}
\alias{parse_convergence_groups}
\title{Parse the groups output by the convergence tool into a data frame}
\usage{
parse_convergence_groups(groups_file, verbose, verbose_dir)
}
\arguments{
\item{groups_file}{A path to the \code{...-convergence-groups.txt} output
file created by the convergence tool. This path is returned by
\code{\link{run_convergence}}.}

\item{verbose}{\code{TRUE} or \code{FALSE}. If \code{TRUE} a CSV
representation of the returned data frame is written to the
\code{verbose_dir}.}

\item{verbose_dir}{A directory to write files to when \code{verbose} is
\code{TRUE}.}
}
\value{
A data frame containing the clusters data. Each row represents a
single cluster. The first column contains the number of items in each cluster
and each cell in the remaining columns is a sequence belonging to that
cluster.
}
\description{
An internal function that extracts the relevant tabular data
from the convergence output file and saves it as a data frame for easy
manipulation and reading. Creates meaningful column names and sorts the
groups from largest to smallest.
}
\keyword{internal}

