% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{parse_convergence_network}
\alias{parse_convergence_network}
\title{Parse the network output by the convergence tool into a data frame}
\usage{
parse_convergence_network(network_file, verbose, verbose_dir)
}
\arguments{
\item{network_file}{A path to the \code{...-clone-network.txt} output
file created by the convergence tool. This path is returned by
\code{\link{run_convergence}}.}

\item{verbose}{\code{TRUE} or \code{FALSE}. If \code{TRUE} a CSV
representation of the returned data frame is written to the
\code{verbose_dir}.}

\item{verbose_dir}{A directory to write files to when \code{verbose} is
\code{TRUE}.}
}
\value{
A data frame containg the network data.
}
\description{
An internal function that puts the network file output by the
tool into a data frame. The network file contains information about which
nodes should connect to which other nodes.
}
\details{
The first and second columns of the data capture which nodes
connect to each other. If there is a node in a group (not this data, but in
the other file output by convergence) it should be looked up in this network
data to see what else the node connects to. Some nodes have no connections,
i.e. if they are not found in this network information. The 3rd column in
the network info contains either 'local', 'global', or 'singleton'.
Singletons have no connections, and local should just be a different color
than global connections in the convergence graph.
}
\keyword{internal}

