% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{parse_convergence_results}
\alias{parse_convergence_results}
\title{Parse the results of the convergence tool into a data frame}
\usage{
parse_convergence_results(results_file, verbose, verbose_dir)
}
\arguments{
\item{results_file}{A path to the \code{...-convergence-groups.txt} output
file created by the convergence tool. This path is returned by
\code{\link{run_convergence}}.}
}
\value{
A data frame containing the clusters data. Each row represents a
single cluster. The first column contains the number of items in each cluster
and each cell in the remaining columns is a sequence belonging to that
cluster.
}
\description{
An internal function that extracts the relevant tabular data
from the convergence output file and saves it as a data frame for easy
manipulation and reading.
}
\keyword{internal}

