% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onload_checks.R
\name{parse_path_string}
\alias{parse_path_string}
\title{Parse system PATH into a list}
\usage{
parse_path_string(path, sep = ":")
}
\value{
A vector of characters containing the various paths.
}
\description{
The system \code{PATH} is usually of format
\code{/usr/bin:/usr/sbin:/usr/local/bin...} where each folder is separated
by a \code{:} character. This function takes the path and splits it into
a list of individual paths so that it can be used elsewhere. The function
has been taken from the \code{rappdirs} package at:
https://github.com/hadley/rappdirs.
}
\keyword{internal}

