% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plotting.R
\name{pca_plot}
\alias{pca_plot}
\title{Plot of first two principal components.}
\usage{
pca_plot(d, clust_obj, num_clust, ...)
}
\arguments{
\item{d}{A numeric matrix of data, or an object that can be coerced to
such a matrix (such as a numeric vector or a data frame with all numeric
columns). Note: This should be the same one used to generate 
\code{clust_obj}.}

\item{clust_obj}{A \code{\link{multiClust-class}} object from which to
extract \code{clust_model} based on the argument \code{num_clust}}

\item{num_clust}{An integer. The desired number of clusters to be used. Note:
This integer should fall within the krange used to generate the 
\code{\link{multiClust-class}} object.}

\item{...}{Further arguments to be passed to the \code{\link{plot}} 
function (besides \code{xlab}, \code{ylab}, \code{main}).}
}
\description{
This function plots the instances of \code{data} using the first two
principal components as the x and y axes, respectively. These components are
calculated using \code{\link[stats]{princomp}}. It represents each cluster
with a different color so that one can understand the distribution of the
clusters based on the first two components. It is a wrapper on the
\code{\link[graphics]{plot}} function.
}
\details{
This function is intended to be used to view the shape of a given
  clustering to use for analysis. Principal component analysis looks for the
  way to represent the data such that the most variance is explained using a
  linear combination of the features. See \code{\link[stats]{princomp}} for
  further details on its calculation. Here it is used strictly for its
  ability to represent the data relatively well in just a couple dimensions.  
  One may pass in the \code{k_best} element from the
  \code{\link{multiClust-class}} object for \code{num_clust}, or use a
  different value. This is where other visualizations can be useful.  
  See the \emph{Details} section for the \code{\link{multi_clust}} function
  and view the \emph{TIP} for a suggested workflow.
}
\examples{
# First, create a multiClust object
library(datasets)
iris_cluster <- multi_clust(iris[, 1:4])
# Second, use object with pca_plot
pca_plot(iris[, 1:4], iris_cluster, num_clust = 3)
}
\seealso{
\code{\link{multi_clust}}, \code{\link{wss_plot}}, 
  \code{\link{gap_plot}}, \code{\link{sil_plot}},
  \code{\link{avg_sil_plot}}, \code{\link{clust_boxplot}}
}

