% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{phyloxml_path}
\alias{phyloxml_path}
\title{Create a path to a tempory PhyloXML}
\usage{
phyloxml_path()
}
\value{
A filepath that can be used to save a PhyloXML.
}
\description{
An internal function that creates a temporary directory to hold
  the PhyloXML file for a phylogram. The directory created is only intended
  to hold the PhyloXML, and no other files, because in order for htmlwidgets
  to make the PhyloXML available in the HTML it copies the entire parent
  directory.
}
\keyword{internal}

