% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{phyloxml_via_biopython}
\alias{phyloxml_via_biopython}
\title{Create a PhyloXML using Biopython}
\usage{
phyloxml_via_biopython(ms_alignment, dist, verbose)
}
\arguments{
\item{ms_alignment}{A multiple sequence alignment of class
\emph{MultipleAlignment}, likely created by the \code{\link{msa}}
function.}

\item{verbose}{\code{TRUE} or \code{FALSE}, depending on whether or not the
stdout and stderr of the Python command should be captured by R. As of now,
seems to have no effect.}
}
\value{
A path to the PhyloXML file.
}
\description{
An internal function that utilizes a custom Python script and
Biopython to create a PhyloXML from a multiple sequence alignment.
}
\keyword{internal}

