% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onload_checks.R
\name{pkg_version}
\alias{pkg_version}
\title{Get installed version of an R package}
\usage{
pkg_version(pkg)
}
\arguments{
\item{pkg}{The name of the R package}
}
\value{
If the package is installed, returns the version number as a
  \code{numeric} (for example, \code{3.8}), otherwise returns NULL.
}
\description{
An internal function that retrieves the version of a R package
}
\examples{
\dontrun{
pkg_version('base')
}
}
\keyword{internal}

