% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/onload_checks.R
\name{py_version}
\alias{py_version}
\title{Get Python version}
\usage{
py_version()
}
\value{
If Python is installed the version number will be returned as a
  \code{numeric} (for example, \code{2.7}). If Python is not installed,
  returns \code{NULL}.
}
\description{
Get the major and minor version of the system's Python
  distribution.
}
\details{
This is an internal function for checking to see whether or not
  Python is installed on the system during the \code{.onLoad()} hook.
}
\keyword{internal}

