% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{radial_phylo}
\alias{radial_phylo}
\title{<Add Title>}
\usage{
radial_phylo(df, seqs_col, canvas_size = "auto", font_size = "auto",
  scale = FALSE, browser = FALSE, width = NULL, height = NULL,
  verbose = FALSE)
}
\description{
<Add Description>
}

