% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{radial_phylo_shiny}
\alias{radial_phyloOutput}
\alias{radial_phylo_shiny}
\alias{renderRadial_phylo}
\title{Shiny bindings for radial phylograms}
\usage{
radial_phyloOutput(outputId, width = "100\%", height = "400px")

renderRadial_phylo(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{The output variable to read from.}

\item{width, height}{A valid CSS unit, such as \code{"100\%"},
\code{"600px"}, \code{"auto"}, or a number (which will be coerced to a string
and have \code{"px"} appended to it).}

\item{expr}{An expression that generates a radial phylogram.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using \code{\link{radial_phylo}}
within Shiny applications and interactive Rmd documents.
}

