% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{radial_phyloOutput}
\alias{radial_phyloOutput}
\title{Widget output function for use in Shiny}
\usage{
radial_phyloOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

