\name{receptormarker-package}
\alias{receptormarker-package}
\alias{receptormarker}
\docType{package}
\title{
\packageTitle{receptormarker}
}
\description{
\packageDescription{receptormarker}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{receptormarker}
\packageIndices{receptormarker}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{receptormarker}

Maintainer: \packageMaintainer{receptormarker}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
