% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receptormarker.R
\docType{package}
\name{receptormarker}
\alias{receptormarker}
\alias{receptormarker-package}
\title{Analyze antibody receptor sequences and phenotypic markers.}
\description{
Useful for quick normalization, clustering, and creating phylogenetic trees.
}

