% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{remove_phyloxml_hash}
\alias{remove_phyloxml_hash}
\title{Remove a dedupe hash from a PhyloXML}
\usage{
remove_phyloxml_hash(xml_file, hash)
}
\arguments{
\item{xml_file}{A path to a PhyloXML file.}

\item{hash}{A string to remove in the PhyloXML. In the XML node(s) in which
\code{hash} is found, any characters following the \code{hash} will also be
removed.}
}
\value{
A path to the PhyloXML file the hash removed.
}
\description{
An internaal function that looks for a hash in all \code{<name>}
  nodes of a PhyloXMl and removes the hash and any remaining characters after
  it. Used in conjuction with the function \code{\link{msa}} when its
  argument \code{dedupe_hash} is used.
}
\seealso{
\code{\link{msa}}
}
\keyword{internal}

