% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{remove_phyloxml_labels}
\alias{remove_phyloxml_labels}
\title{Remove sequence labels from PhyloXML}
\usage{
remove_phyloxml_labels(xml_file)
}
\arguments{
\item{xml_file}{A path to a PhyloXML file.}
}
\value{
A path to the PhyloXML file with sequences removed.
}
\description{
An internal function that removes the labels from the phylogram.
}
\seealso{
\code{\link{radial_phylo}}
}
\keyword{internal}

