% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_plotting.R
\name{renderConvergence}
\alias{renderConvergence}
\title{Widget render function for use in Shiny}
\usage{
renderConvergence(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

