% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{renderRadial_phylo}
\alias{renderRadial_phylo}
\title{Widget render function for use in Shiny}
\usage{
renderRadial_phylo(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

