% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{run_convergence}
\alias{run_convergence}
\title{Create an interactive clustering of adaptive repertoire convergence}
\usage{
run_convergence(seqs, verbose, verbose_dir)
}
\arguments{
\item{seqs}{A character vector containing sequences, preferably cleaned by
\code{\link{clean_data}}.}

\item{verbose}{\code{TRUE} or \code{FALSE}, depending on whether or not the
output of the convergence script should be printed and the output file should
be copied to the \code{verbose_dir}.}

\item{verbose_dir}{A directory to write files to when \code{verbose} is
\code{TRUE}.}
}
\value{
A path to the output file containing the convergence groups in
tabular format.
}
\description{
An internal function that utilizes Jacob Glanville's
\code{convergence.pl} script and associated files to find statistical
clustering of adaptive repertoire convergence.
}
\details{
Requires HMMER and blastn libraries. These binaries are included
in the R package, but may present difficulties on non-Unix computers.
}
\keyword{internal}

