% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totwss.R
\name{totwss}
\alias{totwss}
\title{Calculate total within sum of squares for a clustering and distance object}
\usage{
totwss(dist_obj, clusters)
}
\arguments{
\item{dist_obj}{The distance object for the data set that has been
clustered.}

\item{clusters}{The cluster assignments the distance object has. This should
be a vector.}
}
\value{
An integer value indicating the total within sum of squares
}
\description{
This is an internal function that returns the total sum of squares for a
given data set by passing it the distance object and cluster assignments. It
should only be used when the clustering object does not contain this
information (e.g. an \code{hclust} object).
}
\keyword{internal}

