% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{validate_arc}
\alias{validate_arc}
\title{Validate arc size to be used with a radial phylogram}
\usage{
validate_arc(arc)
}
\arguments{
\item{arc}{An arc size in degrees.}
}
\description{
An internal function that raises an error if an arc size is not
an integer between 1 and 359.
}
\keyword{internal}

