% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{validate_canvas_size}
\alias{validate_canvas_size}
\title{Validate a radial phylogram's canvas size}
\usage{
validate_canvas_size(canvas_size)
}
\arguments{
\item{canvas_size}{}
}
\description{
An internal function that creates an error if \code{canvas_size}
is invalid.
}
\keyword{internal}

