% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{validate_color_wheel}
\alias{validate_color_wheel}
\title{Validate the \code{color_wheel} for a phylogram}
\usage{
validate_color_wheel(color_wheel)
}
\arguments{
\item{color_wheel}{The \code{color_wheel} argument supplied to make the
radial phylogram.}
}
\description{
An internal function that raises an error if \code{color_wheel}
is not a named character vectors or if the hex codes are not valid.
}
\keyword{internal}

