% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_validation.R
\name{validate_convergence_clust}
\alias{validate_convergence_clust}
\title{Validate than an object is of class \emph{convergenceGroups}}
\usage{
validate_convergence_clust(convergence_obj)
}
\arguments{
\item{convergence_obj}{An item to be checked for class membership.}
}
\description{
An internal function that raises an error if the argument is not
of class \emph{convergenceGroups} or if the \code{groups} are empty (i.e.
if there are no convergence groups).
}
\keyword{internal}

