% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{validate_d_seqs}
\alias{validate_d_seqs}
\title{Validate the data, sequence column, and sequences for a phylogram}
\usage{
validate_d_seqs(d, seqs_col)
}
\arguments{
\item{d}{A data frame or character vector containing protein sequences.}

\item{seqs_col}{If \code{d} is a data frame, either an integer corresponding
to a column index or a string corresponding to a column name in \code{d}
that contains the sequences. If \code{d} is a character vector,
\code{seqs_col} should be \code{NULL}.}
}
\description{
An internal function that raises an error if the arguments are
invalid, the data cannot be subset using the provided sequence column, or the
sequences extracted are invalid.
}
\keyword{internal}

