% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{validate_distance}
\alias{validate_distance}
\title{Validate the distance calculator to use on sequence alignment}
\usage{
validate_distance(dist)
}
\arguments{
\item{dist}{A string representing the distance matrix to use for distance
calculations.}
}
\description{
The distance calculator that BioPython creates uses any one of a
number of distance matrices. This function validates that the user input
a single one of these matrices to use for distance calculation.
}
\keyword{internal}

