% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{validate_font_size}
\alias{validate_font_size}
\title{Validate a font size to be used with a radial phylogram}
\usage{
validate_font_size(font_size)
}
\arguments{
\item{font_size}{A font size}
}
\description{
An internal function that raises an error if a font size is not
an integer between 1 and 99.
}
\keyword{internal}

