% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_validation.R
\name{validate_k_range}
\alias{validate_k_range}
\title{Validate the krange parameter of the multi_clust function}
\usage{
validate_k_range(range)
}
\arguments{
\item{n_range}{An item to be checked to make sure it is a valid range.}
}
\description{
An internal function that raises an error if the argument is not
a positive, non-duplicate range of integers beginning > 1. It is sorted into
ascending order.
}
\keyword{internal}

