% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_validation.R
\name{validate_multi_clust_algorithm}
\alias{validate_multi_clust_algorithm}
\title{Validate the 'algorithm' parameter of the multi_clust function}
\usage{
validate_multi_clust_algorithm(algorithm)
}
\arguments{
\item{algorithm}{The \code{algorithm} passed to \code{\link{multi_clust}}.}
}
\description{
An internal function that raises an error if the argument is not
one of "kmeans" or "hclust".
}
\keyword{internal}

