% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_validation.R
\name{validate_multi_clust_method}
\alias{validate_multi_clust_method}
\title{Validate the 'method' parameter of the multi_clust function}
\usage{
validate_multi_clust_method(method)
}
\arguments{
\item{method}{The \code{method} passed to \code{\link{multi_clust}}.}
}
\description{
An internal function that raises an error if the argument is not
one of "silhouette" or "exhaustive".
}
\keyword{internal}

