% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_validation.R
\name{validate_not_null}
\alias{validate_not_null}
\title{Validate that a list of arguments are not \code{NULL}}
\usage{
validate_not_null(arg_list)
}
\arguments{
\item{arg_list}{A named list containing the variables to check. The names in
the list should correspond to the variable names. For example,
\code{list(param=param, datum=datum)}.}
}
\description{
An internal function that raises an error if any of the items in
the list is \code{NULL}.
}
\keyword{internal}

