% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arg_validation.R
\name{validate_num_data}
\alias{validate_num_data}
\title{Validate that the arg is either a numeric data.frame or matrix}
\usage{
validate_num_data(d)
}
\arguments{
\item{d}{A \emph{data.frame} or \emph{matrix} whose class the function will
confirm.}
}
\description{
An internal function that raises an error if the argument is not
either a \emph{data.frame} or \emph{matrix}. Also, if all columns are not
numeric, it will raise an error.
}
\keyword{internal}

