% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arg_validation.R
\name{validate_pos_num}
\alias{validate_pos_num}
\title{Validate that an argument contains positive integers}
\usage{
validate_pos_num(n)
}
\arguments{
\item{n}{A named list of items to be checked.}
}
\description{
An internal function that raises an error if the argument does
not contain positive integers.
}
\keyword{internal}

