% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{validate_ring_thickness}
\alias{validate_ring_thickness}
\title{Validate a ring thickness to be used with a radial phylogram}
\usage{
validate_ring_thickness(ring_thickness)
}
\arguments{
\item{ring_thickness}{A ring thickness}
}
\description{
An internal function that raises an error if a ring
thickness is not an integer between 1 and 99.
}
\keyword{internal}

