% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{validate_rings}
\alias{validate_rings}
\title{Validate the user-supplied argument \code{rings} for a phylogram}
\usage{
validate_rings(rings, d)
}
\arguments{
\item{rings}{A user-supplied named character vector representing rings to add
to a phylogram.}

\item{d}{A data frame or character vector containing protein sequences.}
}
\description{
An internal function that raises an error if \code{rings} is not
a named character vector, if its names do not appear in \code{d} as column
names, if \code{d} is not a data frame, and more.
}
\keyword{internal}

