% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arg_validation.R
\name{validate_sequences}
\alias{validate_sequences}
\title{Validate protein or DNA sequences}
\usage{
validate_sequences(seqs)
}
\arguments{
\item{seqs}{A character vector of sequences.}
}
\description{
An internal function that creates an error if sequences contain
characters outisde the alphabet.
}
\keyword{internal}

