% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_validation.R
\name{validate_single_pos_num}
\alias{validate_single_pos_num}
\title{Validate than an argument is a single, positive integer}
\usage{
validate_single_pos_num(n)
}
\arguments{
\item{n}{An item to be checked.}
}
\description{
An internal function that raises an error if the argument is not
a positive integer.
}
\keyword{internal}

