% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_clustering.R
\name{validate_sort_range}
\alias{validate_sort_range}
\title{Validate and sort an argument that is a range of integers}
\usage{
validate_sort_range(n_range)
}
\arguments{
\item{n_range}{An item to be checked to make sure it is a valid range.}
}
\description{
An internal function that raises an error if the argument is not
a positive, non-duplicate range of integers beginning > 1. It is sorted into
ascending order.
}
\keyword{internal}

