% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{validate_tree_margin}
\alias{validate_tree_margin}
\title{Validate the tree margin to be used with a radial phylogram}
\usage{
validate_tree_margin(tree_margin)
}
\arguments{
\item{tree_margin}{A tree margin}
}
\description{
An internal function that raises an error if a tree margin
is not a number between 0 and 1.
}
\keyword{internal}

