% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arg_validation.R
\name{validate_true_false}
\alias{validate_true_false}
\title{Validate that a list of arguments are \code{TRUE} or \code{FALSE}}
\usage{
validate_true_false(arg_list)
}
\arguments{
\item{arg_list}{A named list containing the variables to check. The names in
the list should correspond to the variable names. For example,
\code{list(param=param, datum=datum)}.}
}
\description{
An internal function that raises an error if any of the items in
the list is not either \code{TRUE} or \code{FALSE}.
}
\keyword{internal}

