% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_plotting.R
\name{wss_plot}
\alias{wss_plot}
\title{Plot of total within sum of squares for numbers of clusters}
\usage{
wss_plot(clust_obj, optimal = FALSE, ...)
}
\arguments{
\item{clust_obj}{A \code{\link{multiClust}} object from which to extract
\code{wss}.}

\item{optimal}{Logical. If \code{TRUE}, the optimal number of clusters as
extracted from \code{clust_obj}, based on average silhouette width, is
circled in the plot.}

\item{...}{Further arguments to be passed to the \code{\link{plot}}
  function (besides \code{xlab}, \code{ylab}, \code{main}).}
}
\description{
This function plots within sum of squares. It provides a wrapper on
\code{\link[graphics]{plot}} by extracting \code{wss} from \code{clust_obj}
and plotting it.
}
\examples{
# First, create a multiClust object
library(datasets)
iris_cluster <- multiClust(iris[, 1:4])
# Second, use object with wss_plot
wss_plot(iris_cluster, optimal = TRUE)
}
\seealso{
\code{\link[graphics]{plot}}, \code{\link{multiClust}}
}

