% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{add_phylo_outer_rings}
\alias{add_phylo_outer_rings}
\title{Add rings to a PhyloXML}
\usage{
add_phylo_outer_rings(xml_file, seqs, d_clean, seqs_col, rings, condensed)
}
\arguments{
\item{xml_file}{A path to a PhyloXML file.}

\item{seqs}{A character vector containing sequences, preferably cleaned by
\code{\link{clean_data}}.}

\item{d_clean}{The cleaned data frame from which the \code{seqs} were
extracted, both preferably cleaned by \code{\link{clean_data}}.}

\item{seqs_col}{Either an integer corresponding to a column index or a string
corresponding to a column name in d that contains the sequences.}

\item{rings}{A named character vector that can be used to create outer-ring
annotations on the radial phylogram. The names of the vector must correspond
to column names in the data.frame \code{d}, and the values should
correspond to a desired value in each column which should be annotated
on the ring. For example, \code{c(FOXP3=1, species="human")} will create
two outer rings, the first of which will be colored whenever the column
"FOXP3" is 1 and the second of which will be colored whenever the column
"species" is "human". Annotations occur on a per-sequence basis when the
PhyloXML represents a non-condensed phylogram. If the PhyloXML represents
a condensed phylogram, annotations occur using individual sequence
populations: if 50\% or more of the cells with a given sequence meet the
current criteria given by \code{rings} then that sequence's ring on the
radial phylogram will be annotated.}

\item{condensed}{\code{TRUE} or \code{FALSE}, depending on whether or not the
PhyloXML represents a condensed phylogram (i.e. a phylogram with only
unique sequences and bars representing clonal expansion).}
}
\value{
A path to the PhyloXML file annotated with rings data.
}
\description{
An internal function that adds a given number of rings to a
condensed or not-condensed radial phylogram. Using the sequences in the
PhyloXML, it examines the provided data \code{d_clean} and adds the XML data
in order to create rings on a radial phylogram whenever the criteria provided
by the argument \code{rings} is found to be true in the data. The rings will
be colored using a preset color scheme.
}
\seealso{
\code{\link{radial_phylo}}
}
\keyword{internal}

