% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plotting.R
\name{avg_sil_plot}
\alias{avg_sil_plot}
\title{Plot average silhouette widths for different numbers of clusters.}
\usage{
avg_sil_plot(clust_obj, optimal = FALSE, ...)
}
\arguments{
\item{clust_obj}{A \code{multiClust} object from which to extract
\code{clust_gap}.}

\item{optimal}{Logical. If \code{TRUE}, the optimal number of clusters as
extracted from \code{clust_obj}, based on average silhouette width, is
circled in the plot.}

\item{...}{Further arguments to be passed to the \code{\link[graphics]{plot}}
function (besides \code{xlab}, \code{ylab}, \code{main}).}
}
\description{
This function plots the average silhouette widths for all numbers of
clusters present in \code{clust_obj}. It provides a wrapper on 
\code{\link[graphics]{plot}} by extracting \code{sil_avg} from 
\code{clust_obj} and plotting it.
}
\details{
This function is intended to be used as part of deciding the ideal
  number of clusters to use for analysis. The average silhouette score tells
  one the goodness of a given clustering looking at a given data points 
  dissimilarity within and outside of its cluster. Typically, one looks for a
  peak before which the average silhouette score rises significantly. See
  \code{\link[cluster]{silhouette}} for further details on its calculation.
  Sometimes, the peak in the score still does not contain enough clusters to
  understand the data well. This is where other visualizations can be useful.  
  Choosing \code{optimal = TRUE} will circle the optimal number of clusters
  based on average silhouette width. See the \emph{Details} section for the
  \code{\link{multi_clust}} function and view the \emph{TIP} for a suggested
  workflow.
}
\examples{
# First, create a multiClust object
library(datasets)
iris_cluster <- multi_clust(iris[, 1:4])
# Second, use object with avg_sil_plot
avg_sil_plot(iris_cluster, optimal = TRUE)
}
\seealso{
\code{\link{multi_clust}}, \code{\link{wss_plot}}, 
  \code{\link{gap_plot}}, \code{\link{pca_plot}}, \code{\link{sil_plot}},
  \code{\link{clust_boxplot}}
}

