% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_utilities.R
\name{clean_data}
\alias{clean_data}
\title{Clean data for a phylogram or convergence and extract the sequences}
\usage{
clean_data(d, seqs_col, verbose, verbose_dir, verbose_format)
}
\arguments{
\item{d}{A data frame or character vector containing protein sequences.}

\item{seqs_col}{If \code{d} is a data frame, either an integer corresponding
to a column index or a string corresponding to a column name in \code{d}
that contains the sequences. If \code{d} is a character vector,
\code{seqs_col} should be \code{NULL}.}

\item{verbose}{\code{TRUE} or \code{FALSE}, depending on whether or not the
cleaned sequences should be written to the \code{verbose_dir}.}

\item{verbose_dir}{A directory to write files to when \code{verbose} is
\code{TRUE}.}

\item{verbose_format}{\code{"fasta"} or \code{"txt"}, indicating how the
cleaned sequences should be written to the \code{verbose_dir}. The options
correspond to FASTA format and a simple text file with each sequence on
its own line. If \code{verbose} is \code{FALSE}, nothing will be written.}
}
\value{
A named list containing the cleaned \code{"seqs"} and \code{"d"}.
}
\description{
An internal function that removes rows of data where the
sequence column is \code{NA} or an empty string.
}
\keyword{internal}

