% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_plotting.R
\name{clust_boxplot}
\alias{clust_boxplot}
\title{Plot cluster membership for each feature.}
\usage{
clust_boxplot(d, clust_obj, num_clust, ...)
}
\arguments{
\item{d}{A numeric matrix of data, or an object that can be coerced to
such a matrix (such as a numeric vector or a data frame with all numeric
columns). Note: This should be the same one used to generate
\code{clust_obj}.}

\item{clust_obj}{A \code{multiClust} object from which to extract
\code{clust_model} based on the argument \code{num_clust}}

\item{num_clust}{An integer. The desired number of clusters to be used. Note:
This integer should fall within the krange used to generate the
\code{multiClust} object.}

\item{...}{Further arguments to be passed to the \code{\link{qplot}}
  function (besides \code{xlab}, \code{ylab}).}
}
\description{
This function plots cluster membership for each feature of \code{d} using box
plots. It utlizes \code{facet_wrap} in the \code{\link[ggplot2]{qplot}}
function and the desired \code{clust_model} from the argument
\code{clust_obj} to show the box plots of each feature after using the
\code{\link[reshape2]{melt}} function to get the correct form of \code{d}.
}
\details{
This function is intended to be used to view the shape of a given
  clustering to use for analysis. One can view which features (likely
  phenotypic markers) are found in each cluster and how the values (likely
  relative expression level) are distributed. This function is primarily
  intended to try to uncover what sort of underlying structure a given
  clustering has found in the data.
  One may pass in the \code{k_best} element from the \code{multiClust} object
  for \code{num_clust}, or use a different value. This is where other
  visualizations can be useful.
  See the \emph{Details} section for the \code{\link{multi_clust}} function
  and view the \emph{TIP} for a suggested workflow.

  \emph{NOTE:} If boolean data is passed in for \code{d}, \code{TRUE} should
  be represented with a \code{1} and \code{FALSE} should be represented with
  a \code{0}. Also, all columns should be boolean in this case so that the
  scaling is correct.
}
\examples{
# First, create a multiClust object
library(datasets)
iris_cluster <- multi_clust(iris[, 1:4])
# Second, use object with clust_boxplot
clust_boxplot(iris[, 1:4], iris_cluster, num_clust = 3)
}
\seealso{
\code{\link{multi_clust}}, \code{\link{wss_plot}},
  \code{\link{gap_plot}}, \code{\link{pca_plot}}, \code{\link{sil_plot}},
  \code{\link{avg_sil_plot}}
}

