% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convergence.R
\name{convergence}
\alias{convergence}
\title{Create an interactive convergence network diagram of similar CDRs}
\usage{
convergence(d, seqs_col = NULL, background_color = "#FFFFFF",
  node_shape = "ELLIPSE", border_width = 2, border_color = "#161616",
  node_color = "#0B94B1", node_size = 30, labels = TRUE,
  label_vertical_pos = "middle", label_horizontal_pos = "center",
  edge_width = 3, edge_color = "#2D2D2D", browser = FALSE,
  verbose = FALSE)
}
\arguments{
\item{d}{A data frame or character vector containing protein sequences.}

\item{background_color}{A valid hex color code as a string for the canvas
color.}

\item{node_shape}{A string indicating what shape to use for the nodes. One of
\code{"ELLIPSE"}, \code{"RECTANGLE"}, \code{"TRIANGLE"}, \code{"DIAMOND"},
\code{"HEXAGON"}, \code{"OCTAGON"}, \code{"PARALLELOGRAM"},
\code{"ROUNDRECT"}, \code{"VEE"}.}

\item{border_width}{An integer border width for the nodes.}

\item{border_color}{A valid hex color code as a string for the border color
of the nodes.}

\item{node_color}{A valid hex color code as a string for the node color.}

\item{node_size}{An integer representing node size.}

\item{labels}{A logical indicating whether or not node labels should be
displayed. If \code{TRUE}, labels will be the sequences used for analysis in
the convergence pipeline.}

\item{label_vertical_pos}{A string indicating the vertical position of node
labels. Must be one of \code{"top"}, \code{"middle"}, or \code{"bottom"}.
Ignored if \code{label=FALSE}.}

\item{label_horizontal_pos}{A string indicating the horizontal position of
node labels. Must be one of \code{"left"}, \code{"center"}, or
\code{"right"}. Ignored if \code{label=FALSE}.}

\item{edge_width}{An integer representing the thickness of edges, or the
lines that connect nodes.}

\item{edge_color}{A valid hex color code as a string for the edge lines.}

\item{browser}{\code{TRUE} or \code{FALSE}, depending on whether or not the
visualization should open in a browser window. Only applicable if using
RStudio, which by default will show the visualization in RStudio's Viewer.
If using R from the command line, the visualization will always open in a
browser.}

\item{verbose}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, additional output
is printed to the console and the sequences, resulting clusters (convergence
groups), and XML file for Cytoscape are written to a folder in the working
directory.}
}
\description{
Creates a JavaScript-based network diagram in RStudio or a
browser window. This visualization can help identify clusters of
functionally similar complementary determining regions (CDRs) of antibody or
T-cell receptors.
}
\examples{
data(tcr)  # Packaged data set, a data.frame from a CSV file
tcr_reduced <- tcr[1:50, ]
}

