% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{convergence}
\alias{convergence}
\title{Group adaptive repertoire convergence by paratope hotspots for CDRs}
\usage{
convergence(d, seqs_col = NULL, verbose = FALSE)
}
\arguments{
\item{d}{A data frame or character vector containing protein sequences.}

\item{seqs_col}{If \code{d} is a data frame, either an integer corresponding
to a column index or a string corresponding to a column name in \code{d}
that contains the sequences. If \code{d} is a character vector,
\code{seqs_col} should be \code{NULL}.}

\item{verbose}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, additional output
is printed to the console and the sequences and resulting cluster
(convergence groups) are written to a folder in the working directory.}
}
\value{
An object of class \code{\link{convergenceGroups-class}} that can be 
inspected or used to generate network graphs using
\code{\link{convergence_plot}}.
}
\description{
Generates convergence motifs and establishes a statistical
cutoff to determine convergence groups. Convergence groups can be plotted
using \code{\link{convergence_plot}}.
}
\details{
Inspection of the convergence groups can (and likely should) be
performed by analyzing the data frame contained within the returned object.
See the examples below for more info.
}
\examples{
data(tcr)  # Packaged data set, a data.frame from a CSV file
tcr_reduced <- tcr[1:100, ]
converged <- convergence(tcr_reduced, seqs_col='seqs')
# You can then: View(converged@groups)
}

