% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{convergence}
\alias{convergence}
\title{<Add Title>}
\usage{
convergence(d, seqs_col = NULL, condense = FALSE, rings = NULL,
  canvas_size = "auto", font_size = 12, scale = TRUE, browser = FALSE,
  verbose = FALSE, fast = FALSE)
}
\description{
<Add Description>
}

