% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_plotting.R
\name{convergence_plot_shiny}
\alias{convergenceOutput}
\alias{convergence_plot_shiny}
\alias{renderConvergence}
\title{Shiny bindings for convergence plots}
\usage{
convergenceOutput(outputId, width = "100\%", height = "400px")

renderConvergence(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{width, height}{A valid CSS unit, such as \code{"100\%"},
\code{"600px"}, \code{"auto"}, or a number (which will be coerced to a string
and have \code{"px"} appended to it).}

\item{expr}{An expression that generates a convergence plot.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{outputID}{The output variable to read from.}
}
\description{
Output and render functions for using
\code{\link{convergence_plot}} within Shiny applications and interactive Rmd
documents.
}

