% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_plotting.R
\name{cytoscape_xml}
\alias{cytoscape_xml}
\title{Create a XML file to plot with Cytoscape}
\usage{
cytoscape_xml(convergence_obj, row_num, labels, verbose, verbose_dir)
}
\arguments{
\item{convergence_obj}{An object of class
\code{\link{convergenceGroups-class}}, typically returned by
\code{\link{run_convergence}}.}

\item{row_num}{An integer indicating which row number in 
\code{convergenceGroups@groups} to graph.
Each row should correspond to a single cluster.}

\item{labels}{\code{TRUE} or \code{FALSE}, depending on whether or not node
labels should be written to the XML and therefore displayed in Cytoscape.}

\item{verbose}{\code{TRUE} or \code{FALSE}. If \code{TRUE} the XML file is
copied to the \code{verbose_dir}.}

\item{verbose_dir}{A directory to write files to when \code{verbose} is
\code{TRUE}.}
}
\value{
A path to the Cytoscape XML file.
}
\description{
An internal function that parses the output of the convergence
tool and creates an XML file to represent a specific row (cluster) in the
file.
}
\keyword{internal}

