% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_plotting.R
\name{pca_plot}
\alias{pca_plot}
\title{Plot of first two principal components.}
\usage{
pca_plot(d, clust_obj, num_clust, ...)
}
\arguments{
\item{d}{A numeric matrix of data, or an object that can be coerced to
such a matrix (such as a numeric vector or a data frame with all numeric
columns). Note: This should be the same one used to generate
\code{clust_obj}.}

\item{clust_obj}{A \code{\link{multiClust}} object from which to extract
\code{clust_model} based on the argument \code{num_clust}}

\item{num_clust}{An integer. The desired number of clusters to be used. Note:
This integer should fall within the krange used to generate the
\code{\link{multiClust}} object.}

\item{...}{Further arguments to be passed to the \code{\link{plot}}
  function (besides \code{xlab}, \code{ylab}, \code{main}).}
}
\description{
This function plots the instances of \code{data} using the first two
principal components as the x and y axes, respectively. It represents each
cluster with a different color so that one can understand the distribution of
the clusters based on the first two components. It is a wrapper on the
\code{\link[graphics]{plot}} function.
}
\examples{
# First, create a multiClust object
library(datasets)
iris_cluster <- multiClust(iris[, 1:4])
# Second, use object with pca_plot
pca_plot(iris[, 1:4], iris_cluster, num_clust = 3)
}
\seealso{
\code{\link[graphics]{plot}}, \code{\link{multiClust}},
}

