% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{radial_phylo}
\alias{radial_phylo}
\title{<Add Title>}
\usage{
radial_phylo(d, seqs_col = NULL, canvas_size = "auto", font_size = "auto",
  scale = TRUE, browser = FALSE, verbose = FALSE)
}
\description{
<Add Description>
}

