% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_plotting.R
\name{sil_plot}
\alias{sil_plot}
\title{Plot silhouette scores for a given clustering of data.}
\usage{
sil_plot(clust_obj, num_clust, ...)
}
\arguments{
\item{clust_obj}{A \code{multiClust} object from which to extract
\code{clust_model} based on the argument \code{num_clust}}

\item{num_clust}{An integer. The desired number of clusters to be used. Note:
This integer should fall within the krange used to generate the
\code{multiClust} object.}

\item{...}{Further arguments to be passed to the \code{\link{plot}}
  function (besides \code{xlab}, \code{ylab}, \code{main}).}
}
\description{
This function plots silhouette scores for each cluster in a given clustering
from \code{clust_obj} by extracting a given \code{sil} from it and leveraging
the special way that the \code{\link[graphics]{plot}} function handles a
\code{\link[cluster]{silhouette}} object.
}
\examples{
# First, create a multiClust object
library(datasets)
iris_cluster <- multi_clust(iris[, 1:4])
# Second, use object with sil_plot
sil_plot(iris_cluster, num_clust = 3)
}
\seealso{
\code{\link[graphics]{plot}}, \code{\link{multi_clust}},
\code{\link[cluster]{silhouette}}
}

