% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_phylo.R
\name{validate_canvas_size}
\alias{validate_canvas_size}
\title{Validate a radial phylogram's canvas size}
\usage{
validate_canvas_size(canvas_size)
}
\arguments{
\item{canvas_size}{A \code{canvas_size} from the \code{\link{radial_phylo}}
function.}
}
\description{
An internal function that creates an error if \code{canvas_size}
is invalid.
}
\keyword{internal}

