% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_clustering.R
\name{validate_multi_clust}
\alias{validate_multi_clust}
\title{Validate that an object is of class \emph{multiClust}}
\usage{
validate_multi_clust(clust_obj)
}
\arguments{
\item{clust_obj}{An item to be checked for class membership.}
}
\description{
An internal function that raises an error if the argument is not
of class \emph{multiClust}.
}
\keyword{internal}

